Revert use of PEP 604 style.
PEP 604 allows writing union types as 'X | Y', however in the case of optional arguments to a function, writing the optional type as 'X | None = None' fails to clearly differentiate between an argument where None is a valid value, and when None is a sentinel value. PEP 645 would allow writing optional types as 'X?', though at the time of this commit, the PEP has not been accepted. For clarity, revert to using 'Union[X, Y]' for union types and 'Optional[X]' for optional arguments.