Store COM object references for SAPI and JFW.
Fixes #2

It seems that at some point it was no longer possible to return a new reference to the SAPI5 COM object from a Python property function. The returned COM object would function normally and return a status value, but wouldn't actually generate any speech.
The solution is to store the reference after it is created by the property function so it is only created once, while maintaining the benefit of delayed creation to module load speed.